refactor: move structs to separate file
